Refactor Flink Submission Service
- Updated Dockerfile to remove log-level argument from uvicorn command.
- Removed unused __init__.py file from app package.
- Created new api package with submit.py for handling the /submit endpoint.
- Implemented submit job logic in submit.py, including pre-submit validation and Flink submission.
- Added core package for business logic, including validation and submission pipeline.
- Created submission command builder and executor for Flink job submission.
- Added validation logic for bundle structure, syntax checks, and import checks.
- Removed old validator.py and exceptions.py files, consolidating functionality into new modules.
- Updated main.py to include new API routes and logging configuration.
- Introduced schemas for request and response models using Pydantic.
- Added utility functions for logging and configuration management.
