+++ modules/processing/flink-submission-service/Dockerfile
@@ -26,4 +26,4 @@ COPY app/ ./app/
 # --------------------------------------------------------------------------- #
 # Start uvicorn
 # --------------------------------------------------------------------------- #
-CMD ["python3", "-m", "uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000", "--log-level", "info"]
+CMD ["python3", "-m", "uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
