--- modules/processing/flink-submission-service/app/__init__.py
@@ -1,1 +1,0 @@
-# flink-submission-service/app package
