+++ modules/processing/flink-submission-service/app/api/__init__.py
@@ -1,0 +1,7 @@
+"""
+api/ – HTTP routes and endpoints.
+"""
+
+from .submit import router
+
+__all__ = ["router"]
