+++ modules/processing/flink-submission-service/app/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+core/ – Business logic: validation and submission pipeline.
+"""
