+++ modules/processing/flink-submission-service/app/core/submission/executor.py
@@ -1,50 +1,15 @@
 """
-submitter.py – Build and execute the Flink CLI submission command.
-
-This module is the only place that calls `flink run --detached`. It never touches
-the filesystem or does validation itself; that is the job of validator.py.
+core/submission/executor.py – Execute Flink submission command.
 """
 
 from __future__ import annotations
 
 import subprocess
 
-from app.models import SubmitRequest, SubmitResult, ValidationResult
-from app.parser import extract_job_id
-from app import settings
-
-
-def build_flink_command(request: SubmitRequest) -> list[str]:
-    """
-    Construct the ``flink run --detached`` command tokens.
-
-    Parameters
-    ----------
-    request:
-        The validated submit request.
-
-    Returns
-    -------
-    list[str]
-        Command tokens suitable for ``subprocess.run()``.
-    """
-    cmd: list[str] = [
-        settings.FLINK_BIN,
-        "run",
-        "--detached",
-        "--jobmanager",
-        settings.JOBMANAGER_ADDRESS,
-        "--python",
-        request.entrypoint,
-    ]
-
-    if request.pyfiles_path is not None:
-        cmd += ["--pyFiles", request.pyfiles_path]
-
-    if request.requirements_path is not None:
-        cmd += ["--pyRequirements", request.requirements_path]
-
-    return cmd
+from app.utils import config
+from app.schemas import SubmitRequest, SubmitResult, ValidationResult
+from .command import build_flink_command
+from .parser import extract_job_id
 
 
 def submit_job(
@@ -65,7 +30,7 @@ def submit_job(
     request:
         The submit request.
     validation:
-        The result of ``run_pre_submit_checks()``.
+        The result of pre-submit validation.
 
     Returns
     -------
@@ -91,7 +56,7 @@ def submit_job(
             cmd,
             capture_output=True,
             text=True,
-            timeout=settings.SUBMIT_TIMEOUT,
+            timeout=config.SUBMIT_TIMEOUT,
         )
     except subprocess.TimeoutExpired:
         return SubmitResult(
@@ -100,7 +65,7 @@ def submit_job(
             returncode=-1,
             stdout="",
             stderr=(
-                f"flink run timed out after {settings.SUBMIT_TIMEOUT}s"
+                f"flink run timed out after {config.SUBMIT_TIMEOUT}s"
             ),
             job_id=None,
             validation=validation,
@@ -112,7 +77,7 @@ def submit_job(
             returncode=-1,
             stdout="",
             stderr=(
-                f"Flink binary not found: '{settings.FLINK_BIN}'. "
+                f"Flink binary not found: '{config.FLINK_BIN}'. "
                 "Is Flink installed and on PATH inside the container?"
             ),
             job_id=None,
