+++ modules/processing/flink-submission-service/app/core/validation/__init__.py
@@ -1,0 +1,20 @@
+"""
+core/validation/ – Pre-submit validation pipeline.
+
+Validation steps:
+  1. Bundle path checks (files / directories exist)
+  2. Python syntax check on entrypoint (ast.parse)
+  3. Python syntax check on pyfiles dir (ast.parse)
+  4. Import check on entrypoint (PYTHONPATH injection)
+"""
+
+from .bundle_structure import validate_bundle_structure
+from .import_check import run_import_check
+from .syntax_checks import run_compileall_check, run_py_compile_check
+
+__all__ = [
+    "validate_bundle_structure",
+    "run_py_compile_check",
+    "run_compileall_check",
+    "run_import_check",
+]
