+++ modules/processing/flink-submission-service/app/utils/__init__.py
@@ -1,0 +1,9 @@
+"""
+utils/ – Shared helper utilities for the Flink submission worker.
+"""
+
+from . import config
+from .logger import configure_logging, get_logger
+
+
+__all__ = ["config", "configure_logging", "get_logger"]
